docker-manager/requirements.txt
docker~=6.1.3
pyyaml==6.0.1
Flask==3.0.0